refactor(core): move labeled_text_input into the core widgets
A generic labeled input consumed by the mnemonic widget and the
watch-only import is core UI kit, not labels UI. The labels facade
stops re-exporting it; consumers import the core path.
